Move plugins command into sidekick_core
