+++ sidekick_core/lib/sidekick_core.dart
@@ -19,6 +19,7 @@ export 'package:sidekick_core/src/commands/analyze_command.dart';
 export 'package:sidekick_core/src/commands/dart_command.dart';
 export 'package:sidekick_core/src/commands/flutter_command.dart';
 export 'package:sidekick_core/src/commands/install_global_command.dart';
+export 'package:sidekick_core/src/commands/plugins_command.dart';
 export 'package:sidekick_core/src/dart.dart';
 export 'package:sidekick_core/src/dart_package.dart';
 export 'package:sidekick_core/src/dart_runtime.dart';
